stt-mom-v1-v2-v3-best-ui-docker-cpu/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
pydantic==2.8.2
python-multipart==0.0.9
faster-whisper==1.0.3
numpy==2.0.1
soundfile==0.12.1
PyYAML==6.0.2
requests==2.32.3
